issue #312: types files
