--- src/types/FormComponentProps.ts
@@ -1,7 +1,0 @@
-import { LabelData } from "./LabelData";
-
-export interface FormComponentProps {
-  title: string;
-  labelData: LabelData;
-  setLabelData: React.Dispatch<React.SetStateAction<LabelData>>;
-}
