--- src/types/LabelData.ts
@@ -1,11 +1,0 @@
-export type LabelData = {
-  companyName: string;
-};
-
-export const DEFAULT_LABEL_DATA: LabelData = {
-  companyName: "",
-};
-
-export const TEST_LABEL_DATA: LabelData = {
-  companyName: "Test Company",
-};
